Ensure delay between notification updates & add related config options
This should ensure that Notification Daemons don't reject the update
request for progress bar. If the break duration is less than 100 seconds
there will not be a progress bar (to be handled in another commit).

Now there is `show_progress_bar` and `minimum_update_delay` options in
the config file.

Fixes #2